requirements.txt
numpy
pandas
streamlit
# openai
python-dotenv
mysql-connector-python
openpyxl
openai==0.28